issue 1: Added DatabaseDataSetConsumerAdapter that supports deferred consumption of the database tables.
