+++ lis-dbunit-migration/lis-dbunit-migration-testcontainers/src/main/java/com/link_intersystems/dbunit/migration/testcontainers/TestcontainersMigrationDataSetTransformerFactory.java
@@ -6,8 +6,9 @@ import com.link_intersystems.dbunit.stream.consumer.ChainableDataSetConsumer;
 import com.link_intersystems.dbunit.stream.consumer.DataSetConsumerPipe;
 import com.link_intersystems.dbunit.testcontainers.DBunitJdbcContainer;
 import com.link_intersystems.dbunit.testcontainers.DatabaseContainerSupport;
+import com.link_intersystems.dbunit.testcontainers.consumer.DatabaseDataSetConsumerAdapter;
 import com.link_intersystems.dbunit.testcontainers.consumer.TestContainersLifecycleConsumer;
-import com.link_intersystems.dbunit.testcontainers.consumer.TestContainersMigrationConsumer;
+import com.link_intersystems.dbunit.testcontainers.consumer.DatabaseOperationConsumer;
 import com.link_intersystems.dbunit.testcontainers.pool.RunningContainerPool;
 import com.link_intersystems.dbunit.testcontainers.pool.SingleRunningContainerPool;
 
@@ -39,12 +40,14 @@ public class TestcontainersMigrationDataSetTransformerFactory implements Migrati
 
         MigrationDatabaseCustomizationConsumer databaseCustomizationConsumer = new MigrationDatabaseCustomizationConsumer(databaseMigrationSupport);
 
-        TestContainersMigrationConsumer testContainersMigrationConsumer = new TestContainersMigrationConsumer();
+        DatabaseOperationConsumer testContainersMigrationConsumer = new DatabaseOperationConsumer();
+        DatabaseDataSetConsumerAdapter databaseDataSetConsumerAdapter = new DatabaseDataSetConsumerAdapter();
 
         DataSetConsumerPipe dataSetConsumerPipe = new DataSetConsumerPipe();
         dataSetConsumerPipe.add(testContainersConsumer);
         dataSetConsumerPipe.add(databaseCustomizationConsumer);
         dataSetConsumerPipe.add(testContainersMigrationConsumer);
+        dataSetConsumerPipe.add(databaseDataSetConsumerAdapter);
 
         return dataSetConsumerPipe;
     }
